+++ __tests__/__configs__/jest.config.js
@@ -1,7 +1,7 @@
 const path = require('path');
 
 module.exports = {
-  collectCoverage: true,
+  collectCoverage: false,
   testMatch: ['**/__tests__/**/*.test.js?(x)'],
   coveragePathIgnorePatterns: ['/node_modules/', '/__configs__/', '/src/', '/view/'],
   testPathIgnorePatterns: ['/node_modules/', '/__configs__/', '/src/', '/view/'],