+++ __tests__/controllers/utilities/object.test.js
@@ -1,5 +1,5 @@
 const {
-  hasOwnKey, cloneDeep, mergeDeep, assignDeep, freezeDeep, proxyfyInCaseInsensitiveKey,
+  hasOwnKey, cloneDeep, mergeDeep, assignDeep, freezeDeep, createCaseInsensitiveProxy,
 } = require(`${global.__ROOT__}/controllers/utilities/object`)[Symbol.for('__TEST__')];
 
 
@@ -23,14 +23,20 @@ describe('Utilities: Object', () => {
     const target = [
       ['a'],
       { a: 'a' },
-      { b: [{ a: 'a' }] },
+      { c: [{ b: { a: ['a'] } }] },
     ];
     //
     const test = target.map(obj => cloneDeep(obj));
+
+    // should clone an array deeply
     expect(test[0]).toEqual(target[0]);
     expect(test[0]).not.toBe(target[0]);
+
+    // should clone an object deeply
     expect(test[1]).toEqual(target[1]);
     expect(test[1]).not.toBe(target[1]);
+
+    // should clone an array/object mixture deeply
     expect(test[2]).toEqual(target[2]);
     expect(test[2]).not.toBe(target[2]);
   });
@@ -40,11 +46,13 @@ describe('Utilities: Object', () => {
     const target = { a: 0, b: 1, c: { d: { e: 2, f: 3 }, g: 4 } };
     const source = { c: { d: { e: 5 }, f: 6 } };
     const result = { a: 0, b: 1, c: { d: { e: 5, f: 3 }, f: 6, g: 4 } };
-    //
+
+    // should immutably merge the source into the target deeply
     const test_1 = mergeDeep(target, source);
     expect(test_1).toEqual(result);
     expect(target).not.toEqual(result);
-    //
+
+    // should mutably merge the source into the target deeply
     const test_2 = mergeDeep(target, source, { mutate: true });
     expect(test_2).toEqual(result);
     expect(target).toEqual(result);
@@ -55,12 +63,14 @@ describe('Utilities: Object', () => {
     const target = { a: { b: { c: { d: { e: { f: 0 } } } } } };
     expect(Object.getOwnPropertyDescriptors(target)).toHaveProperty('a.writable', true);
     expect(Object.getOwnPropertyDescriptors(target.a.b.c.d.e)).toHaveProperty('f.writable', true);
-    //
+
+    // should immutably freeze the target deeply
     const test_1 = freezeDeep(target);
     expect(test_1).not.toBe(target);
     expect(Object.getOwnPropertyDescriptors(test_1)).toHaveProperty('a.writable', false);
     expect(Object.getOwnPropertyDescriptors(test_1.a.b.c.d.e)).toHaveProperty('f.writable', false);
-    //
+
+    // should mutably freeze the target deeply
     const test_2 = freezeDeep(target, { mutate: true });
     expect(test_2).toBe(target);
     expect(Object.getOwnPropertyDescriptors(test_2)).toHaveProperty('a.writable', false);
@@ -73,21 +83,24 @@ describe('Utilities: Object', () => {
     const path = 'a[b].c.d[e][f]';
     const value = 'f';
     const result = { a: { b: { c: { d: { e: { f: 'f' } } }, g: 'g' } } };
-    //
+
+    // should immutably assign value by object path in depth
     const test_1 = assignDeep(target, path, value);
     expect(test_1).not.toBe(target);
     expect(test_1).toEqual(result);
-    //
+
+    // should mutably assign value by object path in depth
     const test_2 = assignDeep(target, path, value, { mutate: true });
     expect(test_2).toBe(target);
     expect(test_2).toEqual(result);
   });
 
 
-  test('Fn: proxyfyInCaseInsensitiveKey', () => {
+  test('Fn: createCaseInsensitiveProxy', () => {
     const target = { a: 1 };
-    const test = proxyfyInCaseInsensitiveKey(target);
-    //
+    const test = createCaseInsensitiveProxy(target);
+
+    // should create a case-insensitive proxy for the target
     expect(test).not.toBe(target);
     expect(test).toEqual(target);
     expect(target.A).toBeUndefined();